requirements-dev.txt
curl_cffi
commonX
PyYAML
Pillow>=10.0.0
psutil
pycryptodome
requests
plugin_jm_server